feat: db validation of the receipt, via new rpc, featured prominently in the header
you now know right away if the receipt is not valid. You can always
easily achieve validity with section at bottom. Need to deal with error
handling - receipt is not valid, WHY? I have the informaiton about the
condition that failed validity, but it is not showed the user anywhere
right now